peers: allow access to underlying connection
This is a first iteration for @awlx. I think this does
require further investigation as I am also not very happy
with the handler interface. As soon as we try to send data
to HAProxy we do need a base implementation that tracks
sticktable ids and allows us to define new one and reference
known sticktables. This is required as we always have to send
HAProxy all sticktables we want to send updates for and can't
reference the ones we already got from that instance.